pyproject.toml
[project]
name = "05-mtusi-netology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=6.13.0",
    "huggingface-hub>=1.12.2",
    "openai>=2.33.0",
    "pandas>=2.3.3",
    "protobuf>=7.34.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.7.2",
    "sentencepiece>=0.2.1",
    "torch>=2.11.0",
    "transformers>=5.7.0",
]
